+++ lisiblepng/src/lisiblepng/bitstream.c
@@ -1,13 +1,13 @@
 #include "bitstream.h"
-#include "assert.h"
+#include <lisiblestd/assert.h>
 
 #define MAX(x, y) (((x) > (y)) ? (x) : (y))
 #define MIN(x, y) (((x) < (y)) ? (x) : (y))
 
 void Bitstream_init(Bitstream *bitstream, const uint8_t *data,
                     size_t data_size) {
-  ASSERT(bitstream != NULL);
-  ASSERT(data != NULL);
+  LSTD_ASSERT(bitstream != NULL);
+  LSTD_ASSERT(data != NULL);
   bitstream->data = data;
   bitstream->data_size = data_size;
   bitstream->byte_offset = 0;
@@ -15,7 +15,7 @@ void Bitstream_init(Bitstream *bitstream, const uint8_t *data,
 }
 
 void bitstream_advance(Bitstream *bitstream, size_t bit_count) {
-  ASSERT(bitstream != NULL);
+  LSTD_ASSERT(bitstream != NULL);
   uint8_t current_bit = bitstream->bit_index;
   bitstream->byte_offset =
       bitstream->byte_offset + (current_bit + bit_count) / 8;
@@ -23,14 +23,15 @@ void bitstream_advance(Bitstream *bitstream, size_t bit_count) {
 }
 
 void Bitstream_skip(Bitstream *bitstream, size_t bit_count) {
-  ASSERT(bitstream != NULL);
+  LSTD_ASSERT(bitstream != NULL);
   bitstream_advance(bitstream, bit_count);
 }
 uint16_t Bitstream_next_bits(Bitstream *bitstream, int bit_count) {
-  ASSERT(bitstream != NULL);
-  ASSERT(bit_count <= 16);
-  ASSERT((bitstream->bit_index + bit_count - 1) / 8 + bitstream->byte_offset <
-         bitstream->data_size);
+  LSTD_ASSERT(bitstream != NULL);
+  LSTD_ASSERT(bit_count <= 16);
+  LSTD_ASSERT((bitstream->bit_index + bit_count - 1) / 8 +
+                  bitstream->byte_offset <
+              bitstream->data_size);
 
   uint16_t val = 0;
   int bits_collected = 0;